% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.nces.programs-package.R
\docType{package}
\name{dc.nces.programs-package}
\alias{dc.nces.programs}
\alias{dc.nces.programs-package}
\title{dc.nces.programs: Social Data Commons: Education & Training: Community College STEM Programs Access Scores}
\description{
Allows user to easily get community college STEM programs access scores data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Hanna Charankevich \email{hc2cc@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
